requirements.txt
python-dotenv
streamlit
streamlit-oauth
dropbox
google-api-python-client
oauth2client
openai
tiktoken
chromadb
langchain
unstructured
unstructured[docx,pdf,xlsx]
youtube-transcript-api
pytube
audio-recorder-streamlit
